+++ extensions-core/histogram/src/main/java/org/apache/druid/query/aggregation/histogram/Histogram.java
@@ -32,7 +32,7 @@ public class Histogram
   {
     double[] retVal = new double[breaks.length];
     for (int i = 0; i < breaks.length; ++i) {
-      retVal[i] = (double) breaks[i];
+      retVal[i] = breaks[i];
     }
 
     this.breaks = retVal;
